requirements.txt
streamlit
streamlit-option-menu
pandas
pillow
joblib
scikit-learn
streamlit-folium
matplotlib
openpyxl
msal
python-stdnum
streamlit-aggrid
firebase-admin
plotly
streamlit-extras
pytz
requests
pyarrow